+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Issue 512](https://github.com/aza547/wow-recorder/issues/512) - Fix a bug where the manager would repeatedly retry configuration if the user got the password wrong.
 - Fixed an issue where you could not download a video if the cloud upload setting was disabled.
 - Fix a bug where downloading the same video twice in a row would fail.
+- Improve simultaenous death handling when applying video timeline marks.
 
 ## [5.7.2] - 2024-08-04
 ### Fixed